+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_command_type.py
@@ -8,10 +8,16 @@ from ._validators import validate_client_parameters
 
 
 def cli_storage_data_plane_command(name, operation, client_factory,
-                                   transform=None, table_transformer=None, exception_handler=None):
+                                   transform=None, table_transformer=None, exception_handler=None,
+                                   resource_type=None, min_api=None, max_api=None):
     """ Registers an Azure CLI Storage Data Plane command. These commands always include the
     four parameters which can be used to obtain a storage client: account-name, account-key,
     connection-string, and sas-token. """
+    if resource_type and (min_api or max_api):
+        from azure.cli.core.profiles import supported_api_version
+        if not supported_api_version(resource_type, min_api=min_api, max_api=max_api):
+            return
+
     command = create_command(__name__, name, operation, transform, table_transformer,
                              client_factory, exception_handler=exception_handler)
     # add parameters required to create a storage client