+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -8,8 +8,10 @@ BC:webapp: fix inconsistencies in the output of "az webapp config appsettings de
 webapp: add a new alias of '-i' for "az webapp config container set --docker-custom-image-name"
 webapp: expose 'az webapp log show'
 webapp: expose new arguments from 'az webapp delete' to retain app service plan, metrics or dns registration. 
+webapp: detect a slot setting correctly 
 
 0.1.13 (2017-08-15)
+=======
 webapp: fix an exception when create a new git based linux webapp
 
 0.1.12 (2017-08-11)