+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -2,6 +2,14 @@
 
 Release History
 ===============
+Unreleased
+++++++++++++++++++
+Breaking Change:webapp: fix inconsistencies in the output of "az webapp config appsettings delete/set"
+webapp: add a new alias of '-i' for "az webapp config container set --docker-custom-image-name"
+webapp: expose 'az webapp log show'
+webapp: expose new arguments from 'az webapp delete' to retain app service plan, metrics or dns registration. 
+webapp: detect a slot setting correctly 
+
 0.1.13 (2017-08-15)
 +++++++++++++++++++
 webapp: fix an exception when create a new git based linux webapp